+++ src/main/java/com/ll/demo/rebate/batch/MakeRebateDataJobConfig.java
@@ -7,7 +7,6 @@ import com.ll.demo.order.repository.OrderItemRepository;
 import com.ll.demo.rebate.entity.RebateItem;
 import com.ll.demo.rebate.repository.RebateItemRepository;
 import lombok.RequiredArgsConstructor;
-import lombok.Value;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.JobScope;
@@ -19,6 +18,7 @@ import org.springframework.batch.item.ItemProcessor;
 import org.springframework.batch.item.ItemReader;
 import org.springframework.batch.item.ItemWriter;
 import org.springframework.batch.item.data.builder.RepositoryItemReaderBuilder;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.Sort;
